# didService/src/main/kotlin/com/did/service/didService/requests/UserUpdateRequest.kt
package com.did.service.didService.requests

import javax.validation.constraints.Email

data class UserUpdateRequest(
    var firstName: String?,
    var lastName: String?,

    @field:Email
    var email: String
)